Correct permissions

+++ utils/server-cache.js
@@ -1,5 +1,7 @@
 import { DOMAIN } from "../constants/router/defaults";
 import { ROUTE_TAGS } from "../core/components/pages/List/constants";
+import ls from "./storage/ls";
+import puppy from "./puppy";
 
 const p = process.env.NODE_ENV === "production" ? "PRODUCTION" : "DEVELOPMENT";
 
@@ -8,14 +10,16 @@ export const invalidate = url => {
   // force cache clear on Next.js server
   fetch(url + "?force=true", { method: "get" }).then(() => {
     // clear CloudFlare caches
-    // process.env.NODE_ENV === "production"
+    if (process.env.NODE_ENV !== "production") return;
+    if (!ls.getItem("token")) return;
     puppy({
       url: `${DOMAIN.PROTOCOL[p] + DOMAIN.API[p]}/admin/cache`,
       method: "DELETE",
       headers: {
         Authorization: "JWT " + ls.getItem("token"),
+        "Content-Type": "application/json",
       },
-      data: {
+      body: {
         files: [url],
       },
     });
